Rename azure.ts to blob.ts
